requirements.txt
httpx==0.27.2
python-dotenv==1.0.1
pytest==8.3.4
